refactor: api group for notification in user center
